+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+### NEXT (YYYY-MM-DD)
+
 ### 1.5.2 (2026-01-09)
 - Do not call curl_close on PHP 8.0+ (#875, @dmaicher)